+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/RedisEntireLookupDao.java
@@ -59,7 +59,7 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
     public List<E> lookup(PagingInfo pagingInfo) throws DaoException {
         try {
             int beginIndex = pagingInfo.getPage() * pagingInfo.getRows();
-            int endIndex = beginIndex + pagingInfo.getRows() - 1;
+            int endIndex = beginIndex + pagingInfo.getRows();
             List<E> es = internalLookup();
             return es.subList(beginIndex, Math.min(es.size(), endIndex));
         } catch (Exception e) {